requirements.txt
aioquic==1.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.3
Flask==3.1.0
h11==0.16.0
h2==4.1.0
hpack==4.1.0
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==3.0.2
mitmproxy==12.1.1
mitmproxy_linux==0.12.7
mitmproxy_rs==0.12.7
msgpack==1.1.0
NetfilterQueue==1.1.0
netifaces==0.11.0
passlib==1.7.4
psutil==7.0.0
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pylsqpack==0.3.22
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyperclip==1.9.0
python-nmap==0.7.1
PyYAML==6.0.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scapy==2.6.1
service-identity==24.2.0
sortedcontainers==2.4.0
tornado==6.5
typing_extensions==4.14.0
urwid==2.6.16
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
